requirements.txt
django
djangorestframework
channels
daphne
psycopg2-binary
